feat: added the logic and types for trackedFields in order to save data according to ERD
