+++ src/index.ts
@@ -4,7 +4,7 @@ import type { ConfigurationDB } from './interfaces/database/ConfigurationDB';
 import type { EventHistoryDB } from './interfaces/database/EventHistoryDB';
 import type { PgQueryConfig } from './builders/utils';
 import type { EvaluationDB } from './interfaces/database/EvaluationDB';
-import type { RawHistoryDB } from './interfaces/database/RawHistoryDB';
+import type { RawHistoryDB, trackedFields, QuarantineRecord } from './interfaces/database/RawHistoryDB';
 import { CreateDatabaseManager, type DatabaseManagerInstance, type ManagerConfig } from './services/dbManager';
 import { LoggerService } from './services/logger';
 import { RedisService } from './services/redis';
@@ -20,4 +20,6 @@ export {
   type EventHistoryDB,
   type EvaluationDB,
   type RawHistoryDB,
+  type trackedFields,
+  type QuarantineRecord,
 };
